:construction:

+++ public/js/application.js
@@ -659,22 +659,24 @@ app.controller('productCreateController', function($scope, $http, Upload) {
     };
 
     $scope.deleteImage = function(index) {
-        if(typeof $scope.product_data.images[index] !== 'object') {
-            $http({
-                method: 'DELETE',
-                url: '/api/products/images?id=' + $scope.product_data.images[index].split('?id=')[1]
-            }).success(function(data, status, headers, config) {
-                console.log(data);
-                if(data.data.result === 'success') {
-                    $scope.product_data.images.splice(index, 1);
-                } else {
-                    console.log(data.data.message);
-                }
-            }).error(function(data, status, headers, config) {
-                console.log(status);
-            });
-        } else {
-            $scope.product_data.images.splice(index, 1);
+        if(confirm('画像を削除しますか？')) {
+            if(typeof $scope.product_data.images[index] !== 'object') {
+                $http({
+                    method: 'DELETE',
+                    url: '/api/products/images?id=' + $scope.product_data.images[index].split('?id=')[1]
+                }).success(function(data, status, headers, config) {
+                    console.log(data);
+                    if(data.data.result === 'success') {
+                        $scope.product_data.images.splice(index, 1);
+                    } else {
+                        console.log(data.data.message);
+                    }
+                }).error(function(data, status, headers, config) {
+                    console.log(status);
+                });
+            } else {
+                $scope.product_data.images.splice(index, 1);
+            }
         }
     };
